backend/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
flask-jwt-extended
werkzeug
python-dotenv    
redis==5.0.1